Fix 'list index out of range' error for capacity variables
The IPCC table does not include 'capacity','cumulative capacity', and 'capacity additions' for all technologies. If the script is trying to write a non-existent variable in the table, 'the 'list index out of range' error will be raised. The 'if' statement makes sure a variable is in the table before attempting to write it.